vessence/requirements.txt
google-adk
chromadb
discord.py
python-dotenv
httpx
aiosqlite
pydantic
edge-tts
